ChocoShop + Cart refactor. Added cart item amounts
