+++ src/components/header/index.js
@@ -4,7 +4,6 @@ import { Container, Menu, MenuList, MenuListItem } from './styles/header'
 export default function Header({ ref, children, ...restProps }) {
   const headerDom = useRef('');
   
-  
   useEffect(() => {
     window.addEventListener('scroll', function() {
       const elem = headerDom.current;
@@ -12,8 +11,12 @@ export default function Header({ ref, children, ...restProps }) {
     });
   }, []);
 
-  
   return <Container ref={ headerDom } { ...restProps }>{ children }</Container>
+};
+
+const handleClick = (e) => {
+  e.preventDefault();
+  window.scrollTo(0, 0);
 };
 
 Header.Menu = function HeaderMenu({ children, ...restProps }) {
@@ -25,5 +28,5 @@ Header.MenuList = function HeaderMenuList({children, ...restProps}) {
 };
 
 Header.MenuListItem = function HeaderMenuListItem({children, ...restProps}) {
-  return <MenuListItem { ...restProps }>{ children }</MenuListItem>
+  return <MenuListItem onClick={handleClick} { ...restProps }>{ children }</MenuListItem>
 };
